gemini-folder-organizerv2/requirements.txt
google-generativeai
python-docx
openpyxl
pymupdf
pillow
python-dotenv
